Teach halberd how to fetch a list of resources from the server
Because the list of resources can vary from cluster to cluster (and
can change even on the same cluster), we can't rely on a static
embedded list of resource types. This allows Halberd to fetch a list
of API resources from the server when requested via the `--update` or
`--update-only` options.